+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@
     using the given mapper function.
 - Bug Fixes
   - Immutables integration doesn't respect @Value.Default for primitives that are nulled in the db
+- Beta API Changes
+  - add GenericTypes.box
+  - QualifiedType: rename mapType -> flatMapType, add a proper mapType
 
 # 3.8.2
 - Improvements